+++ src/Curl/CurlException.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Vjik\TelegramBot\Api\Curl;
+
+use RuntimeException;
+
+/**
+ * @internal
+ */
+final class CurlException extends RuntimeException {}