+++ src/Transport/PsrTransport.php
@@ -5,10 +5,12 @@ declare(strict_types=1);
 namespace Vjik\TelegramBot\Api\Transport;
 
 use Http\Message\MultipartStream\MultipartStreamBuilder;
+use Psr\Http\Client\ClientExceptionInterface;
 use Psr\Http\Client\ClientInterface;
 use Psr\Http\Message\RequestFactoryInterface;
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\StreamFactoryInterface;
+use Psr\Http\Message\StreamInterface;
 use Vjik\TelegramBot\Api\Type\InputFile;
 
 use function is_scalar;
@@ -47,6 +49,51 @@ final readonly class PsrTransport implements TransportInterface
             $response->getStatusCode(),
             $body->getContents(),
         );
+    }
+
+    public function downloadFile(string $url): string
+    {
+        return $this->internalDownload($url)->getContents();
+    }
+
+    public function downloadFileTo(string $url, string $savePath): void
+    {
+        $body = $this->internalDownload($url);
+
+        $content = $body->detach();
+        $content ??= $body->getContents();
+
+        set_error_handler(
+            static function (int $errorNumber, string $errorString): bool {
+                throw new SaveFileException($errorString);
+            },
+        );
+        try {
+            file_put_contents($savePath, $content);
+        } finally {
+            restore_error_handler();
+        }
+    }
+
+    /**
+     * @throws DownloadFileException
+     */
+    private function internalDownload(string $url): StreamInterface
+    {
+        $request = $this->requestFactory->createRequest('GET', $url);
+
+        try {
+            $response = $this->client->sendRequest($request);
+        } catch (ClientExceptionInterface $exception) {
+            throw new DownloadFileException($exception->getMessage(), previous: $exception);
+        }
+
+        $body = $response->getBody();
+        if ($body->isSeekable()) {
+            $body->rewind();
+        }
+
+        return $body;
     }
 
     /**