+++ src/Transport/SaveFileException.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Vjik\TelegramBot\Api\Transport;
+
+use RuntimeException;
+
+/**
+ * @api
+ */
+final class SaveFileException extends RuntimeException {}